CHROMIUM: mm: lower oom dump burst limit to 1
Dumping tasks during oom seems to cause noticeable UI jank.  The kernel
has a mechanism for disabling the dump altogether, but the dump can
still be useful to see in feedback reports. So, this lowers the
default ratelimit burst from 10 to 1 so that we still get one report
of what was consuming memory when we get OOMs.

BUG=chromium:702707
TEST=induce an OOM and check for UI jank

Change-Id: Ia0e73640ad706d8d4f0d6593a4d3ec36e5cd8728
Reviewed-on: https://chromium-review.googlesource.com/501310
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>

Conflicts:
	mm/oom_kill.c

[rebase412(groeck): Context changes]

[rebase419(groeck): Context changes]

Change-Id: Ia34af13ed019ae43d756726b5b1a5a666bec4030

Signed-off-by: fjrXTR <fajarslebew31@gmail.com>