FROMLIST: remoteproc: core: Invoke subdev callbacks in list order
Subdevices at the beginning of the subdev list should have
higher priority than those at the end of the list. Reverse
traversal of the list causes priority inversion, which can
impact the performance of the device.

For example a device adds the glink, sysmon and ssr subdevs
to its list. During a crash the ssr notification would go
before the glink and sysmon notifications. This can cause a
degraded response when a client driver waits for a response
from the crashed rproc.

Signed-off-by: Siddharth Gupta <sidgup@codeaurora.org>

Bug: 188764827
Link: https://lore.kernel.org/linux-arm-msm/1621292893-18827-1-git-send-email-sidgup@codeaurora.org/
Change-Id: Iee635c0a229fae58923075c50c3aa259e20718c7
Signed-off-by: Siddharth Gupta <quic_sidgup@quicinc.com>
Signed-off-by: Elliot Berman <eberman@codeaurora.org>
Signed-off-by: Divya Sharma <divyash@codeaurora.org>